Add files via upload
The Movie Management System in JavaScript is a simple project designed in JavaScript language using HTML and CSS platform. It is very easy to use the features of this Movie Management System, so even users would not find it difficult to use the system. In the text fields, you simply type or pick any movie of your choice and on the Add Movie button you write the details of the movie. You can easily delete any movies if you want to drop them from the list.